+++ Node/index.js
@@ -6,7 +6,7 @@ import {join} from 'path';
 import cheerio from 'cheerio';
 import xpath from 'xpath';
 import xmldom from 'xmldom';
-import * as JSONPath from 'jsonpath-plus';
+import jsonata from 'jsonata';
 
 const ignoreQuerySupport = true;
 
@@ -72,6 +72,17 @@ function getHttpQuery (cfg = {}) {
       'Content-Type': isJSON || forceJSON ? 'application/json' : resultContentType
     };
 
+    const finish = () => {
+      fileContents = forceJSON ? JSON.stringify(queryResult) : queryResult;
+
+      write(res, 200, responseHeaders, fileContents);
+
+      if (next) {
+        // eslint-disable-next-line node/callback-return -- Not that type
+        next();
+      }
+    };
+
     url = (url.slice(-1) === '/' ? url + 'index.html' : url).replace(/\?.*$/u, '');
     // url = require('url').parse(url).pathname; // Need to strip off request parameters?
     // console.log('url:'+url);
@@ -79,7 +90,7 @@ function getHttpQuery (cfg = {}) {
       responseHeaders['query-content-type'] = resultContentType;
     }
 
-    if (req.headers['query-client-support'] && !req.headers['query-request-xpath1'] && !req.headers['query-request-css3'] && !req.headers['query-full-request']) {
+    if (req.headers['query-client-support'] && !req.headers['query-xpath1'] && !req.headers['query-css3'] && !req.headers['query-full-request']) {
       responseHeaders['query-server-support'] = 'xpath1 css3';
       write(res, 200, responseHeaders, ''); // Don't waste bandwidth if client supports protocol and hasn't asked us to deliver the full document
       // Todo: we should allow delivery of a default payload (e.g., full doc if not specified as requesting empty for feature detection+immediate execution if supported)
@@ -109,26 +120,41 @@ function getHttpQuery (cfg = {}) {
     };
 
     let queryResult;
-    if ((ignoreQuerySupport || clientJSONPathSupport) && req.headers['query-request-jsonpath'] && !req.headers['query-full-request']) {
-      queryResult = JSON.stringify(JSONPath.JSONPath({
-        json: JSON.parse(fileContents.toString('utf8')),
-        path: req.headers['query-request-jsonpath'].trim()
-      }));
-    } else if ((ignoreQuerySupport || clientXPath1Support) && req.headers['query-request-xpath1'] && !req.headers['query-full-request']) {
+    if ((ignoreQuerySupport || clientJSONPathSupport) && req.headers['query-jsonata'] && !req.headers['query-full-request']) {
+      const jsonataExpression = jsonata(
+        req.headers['query-jsonata'].trim()
+      );
+      const bindings = req.headers['query-jsonata-bindings']?.trim();
+      jsonataExpression.evaluate(
+        JSON.parse(fileContents.toString('utf8')),
+        bindings ? JSON.parse(bindings) : {},
+        // eslint-disable-next-line promise/prefer-await-to-callbacks -- jsonata API
+        (error, result) => {
+          if (error) {
+            exitError(res, responseHeaders, error);
+            return;
+          }
+
+          queryResult = JSON.stringify(result);
+          finish();
+        }
+      );
+      return;
+    } else if ((ignoreQuerySupport || clientXPath1Support) && req.headers['query-xpath1'] && !req.headers['query-full-request']) {
       const nodeArrayToSerializedArray = (arr) => {
         return arr.map((node) => {
           return node.toString();
         });
       };
       const doc = new xmldom.DOMParser().parseFromString(String(fileContents));
-      const xpath1Request = req.headers['query-request-xpath1'] && req.headers['query-request-xpath1'].trim(); // || '//b[position() > 1 and position() < 4]'; // || '//b/text()',
+      const xpath1Request = req.headers['query-xpath1'] && req.headers['query-xpath1'].trim(); // || '//b[position() > 1 and position() < 4]'; // || '//b/text()',
       queryResult = xpath.select(xpath1Request, doc);
       queryResult = forceJSON ? nodeArrayToSerializedArray(queryResult) : wrapFragment(nodeArrayToSerializedArray(queryResult).join(''));
-    } else if ((ignoreQuerySupport || clientCSS3Support) && req.headers['query-request-css3'] && !req.headers['query-full-request']) {
+    } else if ((ignoreQuerySupport || clientCSS3Support) && req.headers['query-css3'] && !req.headers['query-full-request']) {
       // Support our own custom :text() and :attr(...) pseudo-classes (todo: do as (two-colon) pseudo-elements instead)
       const $ = cheerio.load(String(fileContents));
       // eslint-disable-next-line unicorn/no-unsafe-regex -- Todo
-      const css3RequestFull = req.headers['query-request-css3'] && req.headers['query-request-css3'].trim().match(/(.*?)(?::(text|attr)\(([^)]*)\))?$/u); // Allow explicit "html" (toString) or "toArray" (or "json")?
+      const css3RequestFull = req.headers['query-css3'] && req.headers['query-css3'].trim().match(/(.*?)(?::(text|attr)\(([^)]*)\))?$/u); // Allow explicit "html" (toString) or "toArray" (or "json")?
       const css3Request = css3RequestFull[1];
       const type = css3RequestFull[2] || (forceJSON ? 'toArray' : 'toString');
       const css3Attr = css3RequestFull[3];
@@ -163,14 +189,7 @@ function getHttpQuery (cfg = {}) {
       queryResult = fileContents.toString('utf8');
     }
 
-    fileContents = forceJSON ? JSON.stringify(queryResult) : queryResult;
-
-    write(res, 200, responseHeaders, fileContents);
-
-    if (next) {
-      // eslint-disable-next-line node/callback-return -- Not that type
-      next();
-    }
+    finish();
   };
 }